project controller refactor
